Fetch cars from API
